\name{cati-package}
\alias{cati-package}
\alias{cati}
\docType{package}
\title{
	Community Assembly by Traits: Individuals and beyond
}

\description{
	A package for community assembly testing in particular using individuals traits values. 
}

\details{
\tabular{ll}{
	Package: \tab Tstats \cr
	Type: \tab Package \cr
	Version: \tab 0.6 \cr
	Date: \tab 2014-04-14 \cr
	License: \tab GPL (>=2) \cr
	Depends: \tab R (>= 3.0.2), ape, nlme, e1071, moments, lattice, rasterVis, hypervolume, mice, ade4 \cr
	}
	This package provide function to calculate T-statistics (Tstats function) and other uni-traits metrics (com.index function) to test community assembly considering the individual and other taxonomic levels. Variance partition (partvar function) and density plot for community studies (plot_dens function) are also available. Finaly, this package include function to summarise index and function to plot standardized effect size of index. 
}

\author{
	Adrien Taudiere; adrien.taudiere@cefe.cnrs.fr
}

\references{
}

\keyword{ package }
\seealso{
}

\examples{
}
