\name{cati-package}
\alias{cati-package}
\alias{cati}
\docType{package}
\title{
	Community Assembly by Traits: Individuals and beyond
}

\description{
	A package to detect and quantify community assembly processes using trait values of individuals or populations, the T-statistics and other metrics, and dedicated null models.
}

\details{
	\tabular{ll}{
		Package: \tab cati \cr
		Type: \tab Package \cr
<<<<<<< HEAD
		Version: \tab 0.96 \cr
		Date: \tab 2014-04-14 \cr
		License: \tab GPL (> = 2) \cr
		Depends: \tab R (> = 3.0.2), nlme, ade4, ape \cr
		Imports: \tab e1071, mice, rasterVis, hypervolume, FD, geometry, vegan \cr
		Suggests: \tab lattice, spacodiR, entropart, fBasics, picante \cr  
	}
	This package provides functions to calculate T-statistics (Tstats function) and other uni-traits metrics (ComIndex function) to test community assembly traits measured on individuals and beyond (e.g. populations, functional groups). Variance partitioning  (partvar function) and density plot (plotDistri function) are also available. Finaly, this package includes functions to summarize community assembly metrics and functions to plot standardized effect size of index.  
	
	Development versions, forum, Darwin finches's dataset, and tutorial are available at https://github.com/adrientaudiere/cati/. The reference manual is available at http://cran.r-project.org/web/packages/cati/cati.pdf and the tutorial at https://github.com/adrientaudiere/cati/blob/Package-cati/tutorials/vignette_Darwin_finches/vignette.pdf. 
=======
		Version: \tab 0.94 \cr
		Date: \tab 2014-04-14 \cr
		License: \tab GPL (> = 2) \cr
		Depends: \tab R (> = 3.0.2), nlme, ade4, ape \cr
		Imports: \tab e1071, mice, rasterVis, hypervolume \cr
		Suggests: \tab lattice, vegan, FD, spacodiR, entropart, fBasics, picante \cr  
	}
	This package provides functions to calculate T-statistics (Tstats function) and other uni-traits metrics (ComIndex function) to test community assembly traits measured on individuals and beyond (e.g. populations, functional groups). Variance partitioning  (partvar function) and density plot (plotDistri function) are also available. Finaly, this package includes functions to summarize community assembly metrics and functions to plot standardized effect size of index.  
	
	Development versions, forum, Darwin finches's dataset, and tutorial are available at http://sourceforge.net/p/cati-r/code/ci/master/tree/. The reference manual is available at http://cran.r-project.org/web/packages/cati/cati.pdf and the tutorial at http://sourceforge.net/p/cati-r/code/ci/master/tree/tutorial/vignettes/vignette.pdf. 
>>>>>>> da4a64abb7861c211f4b13b00d2566b05cc7c20b
}

\author{
	Adrien Taudiere; adrien.taudiere@cefe.cnrs.fr
	Cyrille Violle
}
