\name{cati-package}
\alias{cati-package}
\alias{cati}
\docType{package}
\title{
	Community Assembly by Traits: Individuals and beyond
}

\description{
	A package to detect and quantify community assembly processes using trait values of individuals or populations, the T-statistics and other metrics, and dedicated null models.
}

\details{
\tabular{ll}{
	Package: \tab cati \cr
	Type: \tab Package \cr
	Version: \tab 0.9 \cr
	Date: \tab 2014-04-14 \cr
	License: \tab GPL (> = 2) \cr
	Depends: \tab R (> = 3.0.2), ape, nlme, e1071, moments, lattice, rasterVis, hypervolume, mice, ade4 \cr
	}
	This package provides functions to calculate T-statistics (Tstats function) and other uni-traits metrics (ComIndex function) to test community assembly traits measured on individuals and beyond (e.g. populations, functional groups). Variance partitioning  (partvar function) and density plot (plotDistri function) are also available. Finaly, this package includes functions to summarize community assembly metrics and functions to plot standardized effect size of index. 
}

\author{
	Adrien Taudiere; adrien.taudiere@cefe.cnrs.fr
	Cyrille Violle
}
