\name{cati-package}
\alias{cati-package}
\alias{cati}
\docType{package}
\title{
	Community Assembly by Traits: Individuals and beyond
}

\description{
	A package for community assembly testing in particular using individuals traits values. 
}

\details{
\tabular{ll}{
	Package: \tab Tstats \cr
	Type: \tab Package \cr
	Version: \tab 0.4 \cr
	Date: \tab 2014-02-14 \cr
	License: \tab GPL-3 \cr
	Depends: \tab ape, nlme, e1071, moments, lattice, mice, ade4 ,R(>=3.0.0) \cr
	Suggests: \tab rasterVis \cr
	}
	This package provide function to calculate T-statistics (Tstats function) and other uni-traits metrics (com.index function) to test community assembly considering the individual and other taxonomic levels. Variance partition (partvar function) and density plot for community studies (plot_dens function) are also available. Finaly, this package include function to summarise index and function to plot standardized effect size of index (e.g. plot_ses function). 
}

\author{
	Adrien Taudiere; adrien.taudiere@cefe.cnrs.fr
}

\references{
	Violle, Cyrille, Brian J. Enquist, Brian J. McGill, Lin Jiang, Cecile H. Albert, Catherine Hulshof, Vincent Jung, et Julie Messier. 2012. The return of the variance: intraspecific variability in community ecology. Trends in Ecology & Evolution 27 (4): 244-252. doi:10.1016/j.tree.2011.11.014.
}

\keyword{ package }
\seealso{
}

\examples{
}
