\name{finch.ind}
\docType{data}
\alias{finch.ind}
\alias{ind.plot.finch}
\alias{sp.finch}
\alias{traits.finch}
\alias{.Random.seed}

\title{
	Finch morphological data
}

\source{http://bioquest.org/birdd/morph.php}

\description{
	Individual morphological data for Darwin"s finches. finch is the all data.frame. ind.plot.finch and sp.finch respectively correspond to the Ilsland and the species attribute of each individual. traits.finch is the matrix of traits with four traits in rows and 2677 individuals in columns. 
}

\usage{
	data(finch.ind)
}

\format{A data.frame of 2677 individuals in rows and 14 columns.}

\details{
	See http://bioquest.org/birdd/morph.php for more information on database.
}

\examples{
	data(finch.ind)
}






