\name{finch}
\docType{data}
\alias{finch}
\alias{ind.plot_finch}
\alias{sp.finch}
\alias{traits.finch}

\title{
	Finch morphological data
}

\source{http://bioquest.org/birdd/morph.php}

\description{
	Darwin finches individual morphological data. finch is the all data.frame. ind.plot_finch and sp.finch respectively correspond to the Ilsland and the species attribute of each individual. traits.finch is the traits matrix with four traits in rows and 2677 individuals in columns. 
}

\usage{
	data(finch.ind)
}

\format{A data.frame of 2677 individuals in rows and 14 columns.}

\details{
	See http://bioquest.org/birdd/morph.php for more information on database.
}

\value{	
}

\references{
}


\note{
}

\seealso{
}

\examples{
	data(finch.ind)
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line

